requirements.txt
pandas
numpy
selenium
webdriver-manager
google-cloud-storage
